requirements.txt
Django==4.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.7
django-filter==23.5
psycopg[binary]==3.1.18
python-decouple==3.8
django-environ==0.11.2

